requirements.txt
# Trading Bot - Versioni compatibili Python 3.12
setuptools==68.0.0  # Importante per Python 3.12
wheel==0.41.2

# Dipendenze principali
numpy==1.24.3
pandas==1.5.3  # Versione più stabile per 3.12
yfinance==0.2.28
plotly==5.17.0
streamlit==1.28.0
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0